move app.ts, actions, and models into top-level /core directory
